QLearning added
QLearningCore run the QLearning cycle
QTable contains the terrain structured as a grid of connected points
StaticRandom is an utility to generate random values. It is better than UnityEngine.Random because it generate random numbers depending on the computer's clock time
Inside the QLearningCore there is the Reward function that will change based on the actions that we will let the villager to achieve
The QTable and the network are generics. In the future, we will create an ad-hoc class to perform multiple actions