Performance upgrade of the QLearning
The QTable was using HashSets and this meant that every time it had to iterate to find a specific node. Now it is using a Dictionary that grants the access with O(1) consumption.
The radius of the villager is now being automatically adjusted based on the scale and the distance to the next state. In this way it should detect up to 9 gameobjects around the villager. This reduces the computation cost of the research for the closest state. The use of this radius is necessary even in test mode because there is a bug that causes the villager to often move to locations that are very close to the expected ones(therefore the key in the dictionary will be unrecognized).
It has been tested with 100 villagers on a grid of 100x100 and the FPS were ~38.
The QTable is now local to each villager for testing purposes(it can be changed really fast).