Added States: - AttackVillager - StealResource - CollectResource - ConsumeResource
Made q-learning use the new states instead of vector3

Using advanced state objects instead of just position of the AI agent,
making it possible for the AI agent to perform advanced executions in the
states.

Not using NetworkGrid anymore, because all of the states are connected
with each other and the AI now decides which connection gives the best reward.

reward function moved into the State class to have more customized rewards,
partly decided with NEAT, when implemented.

made reward function in State class abstract & implemented reward into each of the states