Implement crossover, historical markings
Implemented the historical markings to allow the crossover between 2
neural networks;
The historical markings are kept in 2 places. One is inside the NEAT to
keep track and sort them all between the different individuals. The other
is inside each neural network, to keep track to the variations of that
specific neural network

There are still some bugs to be solved. They shouldn't take that long.