2 new states added
added following states:
- Attack Villager
this state teleports the agent to another villager alive and attacks it
att the moment it deals 5.0 damage, which directly kills the opponent.
Later this value might need to be played with even though it works fine
with it being at 5

- Steal Resource
this state teleports the agent to another alive villager, which also have
resources in its inventory, and steals one resource.

added the output nodes 7 and 8 to the NEAT to make it evolve reward
multipliers for these 2 new states.